Cleaned up all affected classes

+++ engine/utils/collections/ArrayUtils.cpp
@@ -15,7 +15,7 @@
 
 namespace Siege::Utils
 {
-void ArrayUtils::AssertIsInBounds(const size_t& index, const size_t& size)
+void ArrayUtils::AssertIsInBounds(const unsigned long& index, const unsigned long& size)
 {
     // TODO: Change this to CC_ASSERT once the logging integration is complete
     assert(IsInBounds(index, size) && "The provided index is out of bounds!");
